% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{Plotting function}
\usage{
myplot(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
This function will plot ggplot2 package's graph
}
\examples{
myplot(new_dataframe)

}
\author{
Leean Jo
}
